requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
pymongo==4.7.2
redis==5.0.4
psycopg2-binary==2.9.9
neo4j==5.19.0
pika==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.2.1
pytest-asyncio==0.23.0